layout/inspector/tests/test_css_property_is_shorthand.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test InspectorUtils::CssPropertyIsShorthand</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  const InspectorUtils = SpecialPowers.InspectorUtils;

  let tests = [
    {
      property: "color",
      expected: false
    },
    {
      property: "background",
      expected: true
    },
    {
      property: "--anything",
      expected: false
    }
  ];

  for (let {property, expected} of tests) {
    let result = InspectorUtils.cssPropertyIsShorthand(property);
    is(result, expected, "checking whether " + property + " is shorthand");
  }

  let sawException = false;
  try {
    let result = InspectorUtils.cssPropertyIsShorthand("nosuchproperty");
  } catch (e) {
    sawException = true;
  }
  ok(sawException, "checking whether nosuchproperty throws");

  </script>
</head>
<body>
<h1>Test InspectorUtils::CssPropertyIsShorthand</h1>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
